Move getDsspResNumForPdbFields and required method to DSSP Parser
